send the presence info down to the client socket
I couldn't find a good way to do this. I tried to subscribe the client channel to the
servers process, but it would give an unhandled message for presence_diff. I don't
know why this specifically works (changing the pubsub to the topic), as I haven't found
out how the handling of presence_diff occurs.